+++ examples/live/readme.cpp
@@ -5,7 +5,6 @@
 #include <databento/live.hpp>
 #include <databento/symbol_map.hpp>
 #include <iostream>
-#include <string>
 #include <thread>
 
 using namespace databento;
@@ -13,8 +12,10 @@ using namespace databento;
 int main() {
   PitSymbolMap symbol_mappings;
 
-  auto client =
-      LiveBuilder{}.SetKeyFromEnv().SetDataset("GLBX.MDP3").BuildThreaded();
+  auto client = LiveBuilder{}
+                    .SetKeyFromEnv()
+                    .SetDataset(Dataset::GlbxMdp3)
+                    .BuildThreaded();
 
   auto handler = [&symbol_mappings](const Record& rec) {
     symbol_mappings.OnRecord(rec);